Add support for "dry runs"
This is a semi-hidden tool, mostly for debugging this script.  It
generates an adoption script that "dry runs" in the target database.

The generated file set will be usable, but there is no way to do
adoption of the schema-source database without manually modifying the
adoption script; the most likely end use would be to test that the
adoption of the schema-source database will be successful without
actually carrying out the adoption, prior to generating the final
set of initial XMigra sources and the real adoption script.